requirements.txt
numpy
matplotlib
ipython
seaborn
pandas
lightgbm
scipy
scikit-learn
plotly
scikit-optimize
yellowbrick
jupyter
jupyter-client
jupyter-console
jupyter-core
tqdm
optuna
kaggle
plotly_express
hdbscan
umap-learn[plot]
jupyterlab
streamlit
hiplot
ipykernel
ipywidgets
jupyterlab-git
jupyter-kite
jupyterlab-kite
pyspark
spotipy